backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
cors==1.0.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-huggingface==0.0.0
langchain-chroma==0.1.0
pysqlite3-binary==3.45.1
sentence-transformers==2.2.2
